Finish auth pages

+++ front-end/src/components/forms/profile-form.jsx
@@ -22,11 +22,16 @@ import { Button } from "@/components/ui/button";
 import { Eye, EyeOff, UserLock, UserPen, ShieldUser } from "lucide-react";
 import { convertDate, fetchFromBackend } from "@/lib/helper";
 import { showToast } from "../features/toaster";
+import { ConfirmDialog } from "../features/confirm-dialog";
+import CheckboxButton from "../buttons/checkbox-button";
 
 export function ProfileForm({ open, setOpen, profile }) {
   const [username, setUsername] = useState(profile.name);
   const [phoneNumber, setPhoneNumber] = useState(profile.phone || "");
   const [showPassword, setShowPassword] = useState(false);
+  const [oldPassword, setOldPassword] = useState("");
+  const [newPassword, setNewPassword] = useState("");
+  const [openDeleteAccountDialog, setOpenDeleteAccountDialog] = useState(false);
 
   const router = useRouter();
   const { setUser } = useUser();
@@ -66,6 +71,14 @@ export function ProfileForm({ open, setOpen, profile }) {
 
   const handleDeleteAccount = async (e) => {
     e.preventDefault();
+    const checkbox = document.getElementById("deleteAccountCheckbox");
+    if (!checkbox.checked) {
+      showToast(
+        "warning",
+        "Anda harus mencentang kotak untuk mengonfirmasi penghapusan akun!"
+      );
+      return null;
+    }
     try {
       const res = await fetchFromBackend("/user", {
         method: "DELETE",
@@ -86,29 +99,68 @@ export function ProfileForm({ open, setOpen, profile }) {
     }
   };
 
-  const handleSavePassword = () => {
-    console.log("Password diubah");
+  const handleSavePassword = async (e) => {
+    e.preventDefault();
+
+    if (oldPassword === newPassword) {
+      showToast(
+        "warning",
+        "Password baru tidak boleh sama dengan password lama!"
+      );
+      return;
+    }
+    if (newPassword.length < 8) {
+      showToast(
+        "warning",
+        "Password baru harus memiliki setidaknya 8 karakter!"
+      );
+      return;
+    }
+    try {
+      const payload = {
+        oldPassword: oldPassword,
+        newPassword: newPassword,
+      };
+
+      const res = await fetchFromBackend("/auth/reset-password", {
+        method: "PUT",
+        body: JSON.stringify(payload),
+      });
+
+      if (!res.ok) {
+        showToast("warning", "Gagal mengubah password!");
+      } else {
+        showToast("success", "Berhasil mengubah password!");
+        setOpen(false);
+      }
+    } catch (error) {
+      showToast(
+        "error",
+        "Terjadi kesalahan, coba lagi nanti!",
+        `${error.message}`
+      );
+    }
   };
 
   return (
-    <Sheet open={open} onOpenChange={setOpen}>
-      <SheetContent side="right" className="max-w-md w-full">
-        <SheetHeader className="border-b-2">
-          <SheetTitle>Pengaturan Akun</SheetTitle>
-        </SheetHeader>
-        <SheetDescription className="hidden" />
-        <div className="p-6 space-y-6 my-auto">
-          <Accordion type="single" collapsible>
-            {/* Informasi Akun */}
-            <AccordionItem
-              value="account-info"
-              className="bg-card px-4 shadow-2xs rounded-2xl mb-3"
-            >
-              <AccordionTrigger className="font-semibold">
-                Informasi Akun
-              </AccordionTrigger>
-              <AccordionContent>
-                <form>
+    <>
+      <Sheet open={open} onOpenChange={setOpen}>
+        <SheetContent side="right" className="max-w-md w-full">
+          <SheetHeader className="border-b-2">
+            <SheetTitle>Pengaturan Akun</SheetTitle>
+          </SheetHeader>
+          <SheetDescription className="hidden" />
+          <div className="p-4 space-y-6 my-auto">
+            <Accordion type="single" collapsible defaultValue="account-info">
+              {/* Informasi Akun */}
+              <AccordionItem
+                value="account-info"
+                className="bg-card px-4 shadow-2xs rounded-2xl mb-3"
+              >
+                <AccordionTrigger className="font-semibold">
+                  Informasi Akun
+                </AccordionTrigger>
+                <AccordionContent>
                   <div className="space-y-4 p-2 text-muted-foreground mb-3">
                     <div className="flex flex-col gap-2">
                       <p className="font-semibold">Username:</p>
@@ -129,7 +181,9 @@ export function ProfileForm({ open, setOpen, profile }) {
                       <p className="font-semibold">No. Telepon:</p>
                       <Input
                         id="phone"
-                        type="text"
+                        type="number"
+                        maxLength="15"
+                        minLength="10"
                         placeholder="Belum ditambahkan"
                         value={phoneNumber}
                         onChange={(e) => setPhoneNumber(e.target.value)}
@@ -158,73 +212,79 @@ export function ProfileForm({ open, setOpen, profile }) {
                     <Button
                       size="lg"
                       className="rounded-lg cursor-pointer"
-                      onClick={handleDeleteAccount}
+                      onClick={() => setOpenDeleteAccountDialog(true)}
                     >
                       Hapus Akun
                       <ShieldUser className="h-5 w-5" />
                     </Button>
                   </div>
-                </form>
-              </AccordionContent>
-            </AccordionItem>
+                </AccordionContent>
+              </AccordionItem>
 
-            {/* Reset Password */}
-            <AccordionItem
-              value="reset-password"
-              className="bg-card px-4 shadow-2xs rounded-2xl mb-6"
-            >
-              <AccordionTrigger className="font-semibold">
-                Ganti Password
-              </AccordionTrigger>
-              <AccordionContent>
-                <form action="">
-                  <div className="sr-only">
-                    <Input
-                      id="email"
-                      type="email"
-                      autoComplete="email"
-                      value={profile.email} // Pre-fill dengan email pengguna
-                      readOnly
-                    />
-                  </div>
+              {/* Reset Password */}
+              <AccordionItem
+                value="reset-password"
+                className="bg-card px-4 shadow-2xs rounded-2xl mb-6"
+              >
+                <AccordionTrigger className="font-semibold">
+                  Ganti Password
+                </AccordionTrigger>
+                <AccordionContent>
                   <div className="space-y-4 p-2 text-muted-foreground">
-                    <div className="relative">
-                      <div className="flex gap-2 mb-3 cursor-pointer hover:text-black">
-                        {showPassword ? (
-                          <span
-                            onClick={() => setShowPassword(false)}
-                            className="flex gap-2"
-                          >
-                            Sembunyikan
-                            <Eye className="relative h-5 w-5" />
-                          </span>
-                        ) : (
-                          <span
-                            onClick={() => setShowPassword(true)}
-                            className="flex gap-2"
-                          >
-                            Tampilkan
-                            <EyeOff className="relative h-5 w-5" />
-                          </span>
-                        )}
-                      </div>
-                      <div className="flex items-center border rounded-md">
+                    <form action="#">
+                      <div className="sr-only">
                         <Input
-                          id="old-password"
-                          type={showPassword ? "text" : "password"}
-                          placeholder="Password lama"
-                          autoComplete="old-password"
+                          id="email"
+                          type="email"
+                          autoComplete="email"
+                          value={profile.email}
+                          readOnly
                         />
                       </div>
-                    </div>
-                    <div className="relative">
-                      <Input
-                        id="new-password"
-                        type={showPassword ? "text" : "password"}
-                        placeholder="Password baru"
-                        autoComplete="new-password"
-                      />
-                    </div>
+                      <div className="relative flex flex-col gap-3 mb-3">
+                        <div className="flex gap-2 cursor-pointer hover:text-black">
+                          {showPassword ? (
+                            <span
+                              onClick={() => setShowPassword(false)}
+                              className="flex gap-2"
+                            >
+                              Sembunyikan
+                              <Eye className="relative h-5 w-5" />
+                            </span>
+                          ) : (
+                          <span
+                              onClick={() => setShowPassword(true)}
+                              className="flex gap-2"
+                            >
+                              Tampilkan
+                              <EyeOff className="relative h-5 w-5" />
+                            </span>
+                          )}
+                        </div>
+                        <div className="flex items-center border rounded-md">
+                          <Input
+                            id="old-password"
+                            type={showPassword ? "text" : "password"}
+                            placeholder="Masukkan password lama"
+                            autoComplete="old-password"
+                            value={oldPassword}
+                            onChange={(e) => setOldPassword(e.target.value)}
+                            required
+                          />
+                        </div>
+                      </div>
+                      <div className="relative">
+                        <Input
+                          id="new-password"
+                          type={showPassword ? "text" : "password"}
+                          placeholder="Masukkan password baru"
+                          autoComplete="new-password"
+                          value={newPassword}
+                          onChange={(e) => setNewPassword(e.target.value)}
+                          required
+                        />
+                      </div>
+                    </form>
                     <Button
                       variant="outline"
                       onClick={handleSavePassword}
@@ -234,12 +294,32 @@ export function ProfileForm({ open, setOpen, profile }) {
                       <UserLock className="h-5 w-5" />
                     </Button>
                   </div>
-                </form>
-              </AccordionContent>
-            </AccordionItem>
-          </Accordion>
-        </div>
-      </SheetContent>
-    </Sheet>
+                </AccordionContent>
+              </AccordionItem>
+            </Accordion>
+          </div>
+        </SheetContent>
+      </Sheet>
+
+      {/* Hapus Akun */}
+
+      <ConfirmDialog
+        open={openDeleteAccountDialog}
+        setOpen={setOpenDeleteAccountDialog}
+        title="Apakah Anda yakin?"
+        description="Akun Anda akan dihapus secara permanen. Tindakan ini tidak dapat dibatalkan."
+        checkbox={
+          openDeleteAccountDialog && (
+            <CheckboxButton
+              id="deleteAccountCheckbox"
+              text="Saya mengerti konsekuensinya."
+            />
+          )
+        }
+        confirmHandle={handleDeleteAccount}
+        confirmText="Hapus akun saya"
+        cancelText="Batalkan"
+      />
+    </>
   );
 }
